likes сщккусе

+++ SENT_HABR/likeapp/views.py
@@ -1,6 +1,4 @@
 import json
-
-from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import HttpResponse
 from django.views import View
 from django.contrib.contenttypes.models import ContentType
@@ -8,32 +6,39 @@ from django.contrib.contenttypes.models import ContentType
 from likeapp.models import LikeDislike
 
 
-class VotesView(LoginRequiredMixin, View):
+class VotesView(View):
     model = None
     vote_type = None
 
     def post(self, request, pk):
-        obj = self.model.objects.get(pk=pk)
-        try:
-            likedislike = LikeDislike.objects.get(content_type=ContentType.objects.get_for_model(obj), object_id=obj.id,
-                                                  user=request.user)
-            if likedislike.vote is not self.vote_type:
-                likedislike.vote = self.vote_type
-                likedislike.save(update_fields=['vote'])
+        if request.user.is_authenticated:
+            obj = self.model.objects.get(pk=pk)
+            try:
+                likedislike = LikeDislike.objects.get(content_type=ContentType.objects.get_for_model(obj),
+                                                      object_id=obj.id,
+                                                      user=request.user)
+                if likedislike.vote is not self.vote_type:
+                    likedislike.vote = self.vote_type
+                    likedislike.save(update_fields=['vote'])
+                    result = True
+                else:
+                    likedislike.delete()
+                    result = False
+            except LikeDislike.DoesNotExist:
+                obj.votes.create(user=request.user, vote=self.vote_type)
                 result = True
-            else:
-                likedislike.delete()
-                result = False
-        except LikeDislike.DoesNotExist:
-            obj.votes.create(user=request.user, vote=self.vote_type)
-            result = True
-
+            return HttpResponse(
+                json.dumps({
+                    "result": result,
+                    "like_count": obj.votes.likes().count(),
+                    "dislike_count": obj.votes.dislikes().count(),
+                    "sum_rating": obj.votes.sum_rating()
+                }),
+                content_type="application/json"
+            )
         return HttpResponse(
             json.dumps({
-                "result": result,
-                "like_count": obj.votes.likes().count(),
-                "dislike_count": obj.votes.dislikes().count(),
-                "sum_rating": obj.votes.sum_rating()
+                "redirect": "/auth/login",
             }),
             content_type="application/json"
         )
